012_Commit03_Khoa: add session attribute representing current logged in account
