--- planny/src/main/java/codeclowns/planny/planny/controller/HomeController.java
@@ -1,38 +1,0 @@
-package codeclowns.planny.planny.controller;
-
-import codeclowns.planny.planny.constant.BasicApiConstant;
-import codeclowns.planny.planny.constant.LoginStatus;
-import codeclowns.planny.planny.data.dto.AccountDto;
-import codeclowns.planny.planny.data.mgt.ResponseObject;
-import codeclowns.planny.planny.service.AccountService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api-public")
-@RequiredArgsConstructor
-public class HomeController {
-    private final AccountService accountService;
-
-    @PostMapping("/login")
-    public ResponseObject<?> doPostLogin(@RequestBody AccountDto accountDto) {
-        var response = new ResponseObject<>();
-        try {
-            var status = accountService.login(accountDto);
-            if (status.equals(LoginStatus.SUCCEED)) {
-                response.setStatus(BasicApiConstant.SUCCEED.getStatus());
-                response.setMessage(status.getStateDescription());
-            } else {
-                response.setStatus(BasicApiConstant.FAILED.getStatus());
-                response.setMessage(status.getStateDescription());
-            }
-        } catch (Exception e) {
-            response.setStatus(BasicApiConstant.ERROR.getStatus());
-            response.setMessage(LoginStatus.ERROR.getStateDescription());
-        }
-        return response;
-    }
-}
